requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
channels>=4.2
channels-redis
psycopg2-binary>=2.9    # Optional if you use PostgreSQL
